Move non-pypi packages to directory in plexpy.
